Refactor Return and Sales UIMessages.
